# compress base64 buffers