--- a/src/bokeh/core/serialization.py
+++ b/src/bokeh/core/serialization.py
@@ -45,6 +45,7 @@ from typing import (
 import numpy as np
 
 # Bokeh imports
+from ..settings import settings
 from ..util.dataclasses import (
     Unspecified,
     dataclass,
@@ -176,11 +177,12 @@ class Buffer:
         return self.data.tobytes() if isinstance(self.data, memoryview) else self.data
 
     def to_compressed_bytes(self) -> bytes:
+        level = settings.compression_level()
         # we do not want the result to be different depending on mtime, since that is
         # irrelevant and also makes things harder to test, but Python 3.11 and 3.12 have
         # bug where using mtime=0 results in the Gzip header OS field varies by platforam
         # instead of getting set to a fixed value of 255. So, for now use mtime=1 instead.
-        return gzip.compress(self.to_bytes(), mtime=1)
+        return gzip.compress(self.to_bytes(), mtime=1, compresslevel=level)
 
     def to_base64(self) -> str:
         return base64.b64encode(self.to_compressed_bytes()).decode("utf-8")