--- a/tests/unit/bokeh/core/test_json_encoder.py
+++ b/tests/unit/bokeh/core/test_json_encoder.py
@@ -17,6 +17,8 @@ import pytest ; pytest
 #-----------------------------------------------------------------------------
 
 # Standard library imports
+import gzip
+from base64 import b64encode
 from math import nan
 
 # Bokeh imports
@@ -32,17 +34,17 @@ from bokeh.core.serialization import Serializer
 #-----------------------------------------------------------------------------
 
 def test_json_encoder():
-    val0 = [None, True, False, -128, -1, 0, 1, 128, nan, {"key_0": b"uvw"}]
+    val0 = [None, True, False, -128, -1, 0, 1, 128, nan]
     rep0 = Serializer().serialize(val0)
 
-    assert rep0.buffers is not None and len(rep0.buffers) == 1
+    assert rep0.buffers is not None and len(rep0.buffers) == 0
 
     assert serialize_json(rep0.content) == """\
-[null,true,false,-128,-1,0,1,128,{"type":"number","value":"nan"},{"type":"map","entries":[["key_0",{"type":"bytes","data":"H4sIAAAAAAACEystKwcARkl/GgMAAAA="}]]}]\
+[null,true,false,-128,-1,0,1,128,{"type":"number","value":"nan"}]\
 """
 
-    assert serialize_json(rep0) == f"""\
-[null,true,false,-128,-1,0,1,128,{{"type":"number","value":"nan"}},{{"type":"map","entries":[["key_0",{{"type":"bytes","data":{{"id":"{rep0.buffers[0].id}"}}}}]]}}]\
+    assert serialize_json(rep0) == """\
+[null,true,false,-128,-1,0,1,128,{"type":"number","value":"nan"}]\
 """
 
     assert serialize_json(rep0.content, pretty=True) == """\
@@ -58,23 +60,11 @@ def test_json_encoder():
   {
     "type": "number",
     "value": "nan"
-  },
-  {
-    "type": "map",
-    "entries": [
-      [
-        "key_0",
-        {
-          "type": "bytes",
-          "data": "H4sIAAAAAAACEystKwcARkl/GgMAAAA="
-        }
-      ]
-    ]
   }
 ]\
 """
 
-    assert serialize_json(rep0, pretty=True) == f"""\
+    assert serialize_json(rep0, pretty=True) == """\
 [
   null,
   true,
@@ -84,25 +74,22 @@ def test_json_encoder():
   0,
   1,
   128,
-  {{
+  {
     "type": "number",
     "value": "nan"
-  }},
-  {{
-    "type": "map",
-    "entries": [
-      [
-        "key_0",
-        {{
-          "type": "bytes",
-          "data": {{
-            "id": "{rep0.buffers[0].id}"
-          }}
-        }}
-      ]
-    ]
-  }}
+  }
 ]\
+"""
+
+def test_json_encoder_bytes():
+    val = {"key": b"uvw"}
+    rep = Serializer().serialize(val)
+    assert rep.buffers is not None and len(rep.buffers) == 1
+
+    encoded_bytes = b64encode(gzip.compress(val["key"], mtime=0)).decode("utf-8")
+
+    assert serialize_json(rep.content) == f"""\
+{{"type":"map","entries":[["key",{{"type":"bytes","data":"{encoded_bytes}"}}]]}}\
 """
 
 def test_json_encoder_dict_no_sort():