--- a/tests/unit/bokeh/core/test_serialization.py
+++ b/tests/unit/bokeh/core/test_serialization.py
@@ -18,8 +18,10 @@ import pytest ; pytest
 
 # Standard library imports
 import datetime as dt
+import gzip
 import sys
 from array import array as TypedArray
+from base64 import b64decode
 from math import inf, nan
 from types import SimpleNamespace
 from typing import Any, Sequence
@@ -309,10 +311,8 @@ class TestSerializer:
         encoder = Serializer(deferred=False)
         val = bytes([0xFF, 0x00, 0x17, 0xFE, 0x00])
         rep = encoder.encode(val)
-        assert rep == BytesRep(
-            type="bytes",
-            data="H4sIAAAAAAACE/vPIP6PAQBXSRCfBQAAAA==",
-        )
+        assert rep["type"] == "bytes"
+        assert gzip.decompress(b64decode(rep["data"])) == val
         assert encoder.buffers == []
 
     def test_typed_array(self) -> None:
@@ -336,15 +336,13 @@ class TestSerializer:
         encoder = Serializer(deferred=False)
         val = TypedArray("i", [0, 1, 2, 3, 4, 5])
         rep = encoder.encode(val)
-        assert rep == TypedArrayRep(
-            type="typed_array",
-            array=BytesRep(
-                type="bytes",
-                data="H4sIAAAAAAACE2NgYGBgBGImIGYGYhYgZgViAD34DIUYAAAA",
-            ),
-            order=sys.byteorder,
-            dtype="int32",
-        )
+        # isinstance not supported for typed dicts, alas
+        # assert isinstance(rep, TypedArrayRep)
+        assert rep.keys() == {"type", "order", "dtype", "array"}
+        assert rep["type"] == "typed_array"
+        assert rep["order"] == sys.byteorder
+        assert rep["dtype"] == "int32"
+        assert rep["array"] == encoder._encode_bytes(memoryview(val))
         assert encoder.buffers == []
 
     def test_ndarray(self) -> None:
@@ -369,16 +367,14 @@ class TestSerializer:
         encoder = Serializer(deferred=False)
         val = np.array([0, 1, 2, 3, 4, 5], dtype="int32")
         rep = encoder.encode(val)
-        assert rep == NDArrayRep(
-            type="ndarray",
-            array=BytesRep(
-                type="bytes",
-                data="H4sIAAAAAAACE2NgYGBgBGImIGYGYhYgZgViAD34DIUYAAAA",
-            ),
-            order=sys.byteorder,
-            shape=[6],
-            dtype="int32",
-        )
+        # isinstance not supported for typed dicts, alas
+        # assert isinstance(rep, NDArrayRep)
+        assert rep.keys() == {"type", "order", "shape", "dtype", "array"}
+        assert rep["type"] == "ndarray"
+        assert rep["order"] == sys.byteorder
+        assert rep["shape"] == [6]
+        assert rep["dtype"] == "int32"
+        assert rep["array"] == encoder._encode_bytes(memoryview(val))
         assert encoder.buffers == []
 
     def test_ndarray_dtypes_shape(self) -> None: