--- a/tests/unit/bokeh/core/test_serialization.py
+++ b/tests/unit/bokeh/core/test_serialization.py
@@ -25,6 +25,7 @@ from base64 import b64decode
 from math import inf, nan
 from types import SimpleNamespace
 from typing import Any, Sequence
+from unittest.mock import patch
 
 # External imports
 import numpy as np
@@ -98,6 +99,74 @@ class SomeDataClass:
     f3: NotRequired[bool | None] = Unspecified
     f4: NotRequired[SomeProps] = Unspecified
     f5: NotRequired[SomeModel] = Unspecified
+
+class TestBuffer:
+
+    def test_ref(self) -> None:
+        buf = Buffer(10, b"\x001\xefabc\x12")
+        assert buf.ref == {'id': 10}
+
+    def test_to_bytes_from_bytes(self) -> None:
+        val = b"\x001\xefabc\x12"
+        buf = Buffer(10, val)
+        ret = buf.to_bytes()
+        assert isinstance(ret, bytes)
+        assert ret == val
+
+    def test_to_bytes_from_memoryview(self) -> None:
+        val = b"\x001\xefabc\x12"
+        view = memoryview(val)
+        buf = Buffer(10, view)
+        ret = buf.to_bytes()
+        assert isinstance(ret, bytes)
+        assert ret == val
+
+    def test_to_compressed_bytes_from_bytes(self) -> None:
+        val = b"\x001\xefabc\x12"
+        buf = Buffer(10, val)
+
+        ret = buf.to_compressed_bytes()
+
+        assert isinstance(ret, bytes)
+
+        # make sure the gzip header "OS" field is *always* set to "unknown" (255)
+        assert ret[9] == 255
+
+        assert gzip.decompress(ret) == val
+
+    def test_to_compressed_bytes_from_memoryview(self) -> None:
+        val = b"\x001\xefabc\x12"
+        view = memoryview(val)
+        buf = Buffer(10, view)
+
+        ret = buf.to_compressed_bytes()
+
+        assert isinstance(ret, bytes)
+
+        # make sure the gzip header "OS" field is *always* set to "unknown" (255)
+        assert ret[9] == 255
+
+        assert gzip.decompress(ret) == val
+
+    @pytest.mark.parametrize("level", range(0, 10))
+    @patch("gzip.compress")
+    def test_to_compressed_compression_level(self, compress, level: int, monkeypatch: pytest.MonkeyPatch) -> None:
+        val = b"\x001\xefabc\x12"
+        buf = Buffer(10, val)
+
+        monkeypatch.setenv("BOKEH_COMPRESSION_LEVEL", f"{level}")
+        buf.to_compressed_bytes()
+
+        compress.assert_called_once_with(val, mtime=1, compresslevel=level)
+
+    def test_to_base64(self):
+        val = b"\x001\xefabc\x12"
+        buf = Buffer(10, val)
+
+        ret = buf.to_base64()
+
+        assert isinstance(ret, str)
+        assert b64decode(ret) == buf.to_compressed_bytes()
 
 class TestSerializer: