… tests/unit/bokeh/test_settings.py