- [x] issues: fixes #14460
- [x] tests added / passed

This PR adds support to transparently compress and decompress any data that would currently be transmitted as a base64 buffer. 

In order to easily support synchronous gzip compress and uncompress on the BokehJS side, [`fflate`](https://github.com/101arrowz/fflate) was added as a dependency.