@bryevdv 2025-05-11
Just noting that the JS tests that *start* from base64 encoded expected values were a bit of a pain to update. I had to run the values though this little python helper and copy/paste the results back into the tests:
```python
import gzip
from base64 import b64encode, b64decode
def refutz(data):
    raw = b64decode(data)
    compressed = gzip.compress(raw, mtime=0)
    return b64encode(compressed)
```
It might be nice to start from actual array/image data in the tests but I guess that's not possible in all cases, e.g. JSON recordings for "cross" tests.