@bryevdv 2025-05-11
Hrm, these tests are passing locally for me:
```
E         -  'array': {'data': 'H4sIAAAAAAACE2NgYGBgBGImIGYGYhYgZgViAD34DIUYAAAA',
E         ?                                 ^
E         +  'array': {'data': 'H4sIAAAAAAAC/2NgYGBgBGImIGYGYhYgZgViAD34DIUYAAAA',
E         ?     
```
Perusing the [Gzip RFC](https://www.rfc-editor.org/rfc/rfc1952.html#section-2.2) it seems as though an "OS" field is encoded in the header, so it looks like those tests will have to be handled a little more carefully...