@bryevdv 2025-05-12
@mattpap the last changes leave some cross tests failing for the above mentioned reasons. What do you want to do here? I simply don't think we can store static JSON in cases like this, given how the gzip header varies by platform.