@philippjfr 2025-05-12
Neat! Would be great to see some performance number in terms of space saving and overhead for compression and decompression. Pretty sure it's worth it but would love to have real numbers.