@bryevdv 2025-05-12
Here's a quick benchmarking script `benchmark.py`

<details>

```python
import argparse
import importlib
from time import time

import pandas as pd

import bokeh
from bokeh.embed import file_html
from bokeh.plotting import figure

# assume 3.8 is local build with compression
compression = "with" if bokeh.__version__.startswith("3.8.0.dev") else "without"

def run(func, N):
    img = func(N)
    t0 = time()
    p = figure(width=400, height=400)
    p.image_rgba(image=[img], x=0, y=0, dw=10, dh=10)
    out = file_html(p)
    t1 = time()
    dt = (t1 - t0)
    size = len(out.encode("utf-8"))
    return (dt, size)

def benchmark(module, course):
    name = module.__name__
    rows = []
    print(f"benchmarking {name} {compression} compression")
    for i in course:
        print(f"running for N={i}")
        result = run(module.func, i)
        rows.append((name, compression, i, *result))
    columns = ("name", "compression", "N", "time", "size")
    return pd.DataFrame(rows, columns=columns)

if __name__ == "__main__":
    parser = argparse.ArgumentParser()

    parser.add_argument('module', type=str, help='module to benchmark')
    parser.add_argument('--start', type=int, help='bench range start', default=50)
    parser.add_argument('--stop', type=int, help='bench range stop', default=3000)
    parser.add_argument('--step', type=int, help='bench range step', default=50)

    args = parser.parse_args()

    module = importlib.import_module(args.module)

    course = range(args.start, args.stop+1, args.step)

    df = benchmark(module, course)

    print(df)

    filename = f"{module.__name__}_{compression}_{args.start}_{args.stop}_{args.step}"
    df.to_csv(filename, index=False)
```

</details>

Run it with a module containing a single function `func(N)` that returns an NxN array of RGBA data, e.g here is `ramp.py` generalized from our `image_rgba.py` example:
```python
import numpy as np

def func(N):
    img = np.empty((N,N), dtype=np.uint32)
    view = img.view(dtype=np.uint8).reshape((N, N, 4))
    for i in range(N):
        for j in range(N):
            view[i, j, 0] = int(i/N*255)
            view[i, j, 1] = 158
            view[i, j, 2] = int(j/N*255)
            view[i, j, 3] = 255
    return img
```

Then run like this passing the module name and any start/stop/step values you want:
```sh
python benchmark.py ramp --step 500
```
that will print t dataframe with the results and also write a CSV with the results that can be combined e.g. for plotting. I'd encourage others to play around with this. 

For now, here are some quick results with the invocation above for with/without compression:

### without compression
```
   name compression     N      time      size
0  ramp     without    50  0.019395     19406
1  ramp     without   550  0.017741   1619408
2  ramp     without  1050  0.036655   5886074
3  ramp     without  1550  0.064069  12819410
4  ramp     without  2050  0.114079  22419410
5  ramp     without  2550  0.186313  34686074
```
### with compression
```
   name compression     N      time    size
0  ramp        with    50  0.038783   12277
1  ramp        with   550  0.041970  145103
2  ramp        with  1050  0.073990  183301
3  ramp        with  1550  0.107363  236729
4  ramp        with  2050  0.151728  305673
5  ramp        with  2550  0.199548  393529
```