@mattpap 2025-05-12
> the last changes leave some cross tests failing for the above mentioned reasons. What do you want to do here? 

I would be surprised if such differences can't be eliminated. This would be a nightmare for computing checksums, among other things. However, if we can't eliminate these, then I would disable compression in cross tests, which should be an option regardless in my opinion.