@bryevdv 2025-05-12
Here are some quick results with a random image (two channels replaced with random ints)
```python
            view[i, j, 0] = int(i/N*255)
            view[i, j, 1] = 158
            view[i, j, 2] = randint(0, 255)
            view[i, j, 3] = randint(0, 255)
```

### without compression
```
   name compression     N      time      size
0  rand     without    50  0.019407     19406
1  rand     without   550  0.017949   1619408
2  rand     without  1050  0.041163   5886074
3  rand     without  1550  0.064462  12819410
4  rand     without  2050  0.105933  22419410
5  rand     without  2550  0.160247  34686074
```

### with compression

```
   name compression     N      time      size
0  rand        with    50  0.032702     16937
1  rand        with   550  0.045025   1124931
2  rand        with  1050  0.215238   4003877
3  rand        with  1550  0.691357   8705253
4  rand        with  2050  1.467675  15167653
5  rand        with  2550  2.431757  23418753
```

As expected, significantly less space savings as well as more expensive in terms of time execution. 

---

So, what is really needed is some image gen functions that produce image that are in between these two extremes, and more representative of real-world images. 

Also note I did not tinker with the default compression levels. I suppose it might be reasonable to make compression level and even use of compression configurable (though I am less sure about that last one).