@bryevdv 2025-05-12
> I would be surprised if such differences can't be eliminated. 

The Gzip RFC specifies (at least) a field to encode OS (e.g. Unix, or... Amiga!) and I didn't see anything in the various APIs analogous to setting `mtime=0` to disregard this. So, maybe not. 

> However, if we can't eliminate these, then I would disable compression in cross tests, which should be an option regardless in my opinion.

I'm on the fence about this, tbh. I think testing for `H4sIAAAA` start of the data would almost be reasonable to auto-detect whether this is compressed data or not, but it's not 100% perfect. I suppose `BytesRep` could get an extra field to flag this...